day_97/server/requirements.txt
flask==3.1.2
flask_sqlalchemy==3.1.1
pyjwt==2.10.1